+++ aiohttp_sqlalchemy/views.py
@@ -1,16 +1,16 @@
 from abc import ABCMeta
 from typing import Any, Optional
 
-from aiohttp.abc import AbstractView
 from aiohttp.web import View
+from aiohttp_things.views import ContextMixin, InstanceMixin, PrimaryKeyMixin
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from aiohttp_sqlalchemy.constants import SA_DEFAULT_KEY
 from aiohttp_sqlalchemy.utils import get_session
 
 
-class SAMixin(AbstractView, metaclass=ABCMeta):
-    """SQLAlchemy view mixin based `aiohttp.abc.AbstractView`."""
+class SAMixin(ContextMixin, metaclass=ABCMeta):
+    """SQLAlchemy class based view mixin."""
 
     sa_session_key: str = SA_DEFAULT_KEY
 
@@ -23,9 +23,23 @@ class SAMixin(AbstractView, metaclass=ABCMeta):
 
 
 class SAModelMixin(SAMixin, metaclass=ABCMeta):
-    """SQLAlchemy single model view mixin based `aiohttp.abc.AbstractView`."""
+    """SQLAlchemy single model class based view mixin."""
 
-    sa_model: Any  # Not all developers use declarative mapping
+    sa_model: Any = None  # Not all developers use declarative mapping
+
+
+class SAInstanceMixin(
+    InstanceMixin,
+    PrimaryKeyMixin,
+    SAModelMixin,
+    metaclass=ABCMeta,
+):
+    """
+    SQLAlchemy single instance class based view mixin.
+
+    :param sa_pk_attr: primary key column or hybrid attribute.
+    """
+    sa_pk_attr: Any = getattr(SAModelMixin.sa_model, 'pk', None)
 
 
 class SABaseView(View, SAMixin):
@@ -36,7 +50,13 @@ class SAModelView(View, SAModelMixin):
     """SQLAlchemy single model class based view."""
 
 
+class SAInstanceView(View, SAInstanceMixin, metaclass=ABCMeta):
+    """SQLAlchemy single instance class based view."""
+
+
 # Synonyms
 SAAbstractView = SAMixin
+SAItemMixin = SAInstanceMixin
+SAItemView = SAInstanceView
 SAOneModelMixin = SAModelMixin
 SAView = SAModelView